requirements.txt
absl-py==0.12.0
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
attrs==21.2.0
billiard==3.6.4.0
blis==0.7.4
boto3==1.17.84
botocore==1.20.84
cachetools==4.2.2
catalogue==2.0.4
celery==5.1.0
certifi==2020.12.5
cfgv==3.3.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorthief==0.2.1
convcolors==1.0.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.23
decorator==4.4.2
demes==0.1.2
deskew==0.10.29
distlib==0.3.2
docopt==0.6.2
easydict==1.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
extcolors==1.0.0
filelock==3.0.12
fire==0.4.0
flatbuffers==1.12
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
grpcio==1.34.1
h5py==3.1.0
identify==2.2.10
idna==2.10
imageio==2.9.0
imutils==0.5.4
jamotools==0.1.10
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
Lasagne @ git+https://github.com/Lasagne/Lasagne.git@61b1ad14d9960a2e20ff2f06ea3f30e81a21441e
lmdb==1.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
msprime==1.0.1
murmurhash==1.0.5
natsort==7.1.1
networkx==2.5.1
newick==1.3.0
nltk==3.6.2
nodeenv==1.6.0
numpy==1.19.5
oauthlib==3.1.0
onnx==1.9.0
onnx-tf==1.8.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pathy==0.5.2
Pillow==8.2.0
pkgconfig==1.5.4
pre-commit==2.13.0
preshed==3.0.5
prompt-toolkit==3.0.18
protobuf==3.17.1
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytorch-transformers==1.2.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.8
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
sacremoses==0.0.45
scikit-image==0.18.1
scipy==1.6.3
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
sqlparse==0.4.1
srsly==2.4.1
svgwrite==1.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
Theano==0.8.2
thinc==8.0.3
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
tskit==0.3.6
typeguard==2.12.1
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.5
vine==5.0.0
virtualenv==20.4.7
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yarg==0.1.9
google-cloud-vision==2.3.1
azure-cognitiveservices-vision-computervision
